more restructuring + readme + package.json
