+++ dbpedia/dbpedia_Classical_musicians_by_century.js
@@ -201,7 +201,7 @@ function fetch2(value2) {
             work: work,
             release: release,
             tag: tag,
-            source_link: value2
+            source_link: value2,
             wikipedia_link: wiki_link,
             wikipedia_pageid: wiki_pageid
         });
