--- dbpedia/dbpedia_Classical_musicians_by_nationality/dbpedia_Classical_musicians_by_nationality_examples.json
@@ -1,1 +1,0 @@
-{"name":"Trey Anastasio","nationality":"American","dateOfBirth":"1964-09-30","dateOfDeath":"","placeOfBirth":"Fort Worth, Texas, U.S.","placeOfDeath":"","instrument":["Vocals, guitar, keyboard, drums, Red Lightsaber"],"psuedonym":[],"work":[],"release":[],"tag":["American classical composers","American classical musicians","American male classical composers"],"link":"http://dbpedia.org/resource/Trey_Anastasio"},{"name":"Chris Freeman (Australian musician)","nationality":"Australian","dateOfBirth":"1950-01-01","dateOfDeath":"1992-01-01","placeOfBirth":"","placeOfDeath":"","instrument":["Guitar , organ, piano, synthesiser, Fairlight CMI"],"psuedonym":[],"work":[],"release":[],"tag":["20th-century classical musicians","Australian classical musicians"],"link":"http://dbpedia.org/resource/Chris_Freeman_(Australian_musician)"},{"name":"Rachael Leahcar","nationality":"Australian","dateOfBirth":"1994-02-22","dateOfDeath":"","placeOfBirth":"Adelaide, South Australia, Australia","placeOfDeath":"","instrument":["Vocals, piano, flute"],"psuedonym":[],"work":[],"release":[],"tag":["Australian classical musicians","Blind classical musicians"],"link":"http://dbpedia.org/resource/Rachael_Leahcar"}
