+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "structured_data",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/MusicConnectionMachine/StructuredData.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/MusicConnectionMachine/StructuredData/issues"
+  },
+  "homepage": "https://github.com/MusicConnectionMachine/StructuredData#readme",
+  "dependencies": {
+    "cheerio": "0.22.0",
+    "fs": "0.0.1-security",
+    "jsesc": "2.4.0",
+    "math-ceil": "1.0.0",
+    "numeral": "2.0.6",
+    "request": "2.81.0",
+    "strsplit": "1.0.0",
+    "system-sleep": "1.3.0"
+  }
+}
